Invoke pkgnmmap with the standard (Build)Requires tags
It is unnecessarily disruptive to define new (Build)RequiresM tags for use in
SFE specs. The existing tags can be used by pkgnmmap. If a given package name
is not in the pkgnmmap database, the behavior of pkgbuild will be as before.
In other words, when pkgbuild encounters a (Build)Requires tag, it checks to
see if the required package is in the pkgnmmap database. If it is, then the
package name given there is used. Otherwise, pkgbuild just uses the package
name given in the spec file.

Thus, the pkgnmmap facility only alters pkgbuild's behavior if (1) mapping.yaml
is present and (2) if at least one package name given in a (Build)Requires is in
the database specified by mapping.yaml.