Handle the package name mapping file like included files in include/
pkgbuild now looks for mapping.yaml in ~/packages/SPECS, and pkgtool copies
it from include/ in the spec files repository to that directory. This is how
files referred to by %include statements in spec files are dealt with.

TODO: $yaml_filename is now defined in both pkgnmmap.pl and pkgtool.pl.
      It should only be defined in one place. Also, the place to look
      for it is currently hardcoded. (The places where pkgtool looks for
      included files aren't defined in pkgtool.pl; I don't bother to
      figure out where they are defined.)